requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

web3==6.15.1
python-dotenv==1.0.1

pydantic==2.7.4
eth-account==0.11.2